Top side fixed, need to finish bottom
